お題のLabelをARの画面上に表示するようにした
